+++ CHANGELOG.md
@@ -1,6 +1,21 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.3.0-DEV] - 2019-02-19
+
+#### Changed
+* increased WebPush lib dependency to 5
+* changed start field of push notifications to mysql varchar type
+* PushNotificationSender now throws errors,
+
+#### Added
+* option to configure push message encryption padding
+
+#### Fixed
+* push notifications not arrive in firefox mobile
+* notification titles not shows
+
+
 ## [0.2.1] - 2019-02-19
 
 #### Fixed